move s3gateway into pkg
