+++ cmd/dump.go
@@ -20,12 +20,14 @@ import (
 	"io"
 	"os"
 
+	"github.com/juicedata/juicefs/pkg/utils"
+
 	"github.com/juicedata/juicefs/pkg/meta"
 	"github.com/urfave/cli/v2"
 )
 
 func dump(ctx *cli.Context) error {
-	setLoggerLevel(ctx)
+	utils.SetLogger(ctx)
 	if ctx.Args().Len() < 1 {
 		return fmt.Errorf("META-URL is needed")
 	}
