+++ cmd/fsck.go
@@ -22,6 +22,7 @@ import (
 	"time"
 
 	"github.com/juicedata/juicefs/pkg/chunk"
+	"github.com/juicedata/juicefs/pkg/common"
 	"github.com/juicedata/juicefs/pkg/meta"
 	"github.com/juicedata/juicefs/pkg/object"
 	osync "github.com/juicedata/juicefs/pkg/sync"
@@ -42,7 +43,7 @@ func checkFlags() *cli.Command {
 }
 
 func fsck(ctx *cli.Context) error {
-	setLoggerLevel(ctx)
+	utils.SetLogger(ctx)
 	if ctx.Args().Len() < 1 {
 		return fmt.Errorf("META-URL is needed")
 	}
@@ -65,7 +66,7 @@ func fsck(ctx *cli.Context) error {
 		CacheSize:  0,
 	}
 
-	blob, err := createStorage(format)
+	blob, err := common.CreateStorage(format)
 	if err != nil {
 		logger.Fatalf("object storage: %s", err)
 	}
