+++ cmd/main.go
@@ -18,13 +18,9 @@ package main
 import (
 	"fmt"
 	"log"
-	"net/http"
 	_ "net/http/pprof"
 	"os"
 	"strings"
-
-	"github.com/google/gops/agent"
-	"github.com/sirupsen/logrus"
 
 	"github.com/juicedata/juicefs/pkg/utils"
 	"github.com/juicedata/juicefs/pkg/version"
@@ -251,32 +247,4 @@ func reorderOptions(app *cli.App, args []string) []string {
 		}
 	}
 	return append(newArgs, others...)
-}
-
-func setupAgent(c *cli.Context) {
-	if !c.Bool("no-agent") {
-		go func() {
-			for port := 6060; port < 6100; port++ {
-				_ = http.ListenAndServe(fmt.Sprintf("127.0.0.1:%d", port), nil)
-			}
-		}()
-		go func() {
-			for port := 6070; port < 6100; port++ {
-				_ = agent.Listen(agent.Options{Addr: fmt.Sprintf("127.0.0.1:%d", port)})
-			}
-		}()
-	}
-}
-
-func setLoggerLevel(c *cli.Context) {
-	if c.Bool("trace") {
-		utils.SetLogLevel(logrus.TraceLevel)
-	} else if c.Bool("verbose") {
-		utils.SetLogLevel(logrus.DebugLevel)
-	} else if c.Bool("quiet") {
-		utils.SetLogLevel(logrus.WarnLevel)
-	} else {
-		utils.SetLogLevel(logrus.InfoLevel)
-	}
-	setupAgent(c)
 }
