+++ pkg/common/common.go
@@ -1,0 +1,120 @@
+/*
+ * JuiceFS, Copyright (C) 2021 Juicedata, Inc.
+ *
+ * This program is free software: you can use, redistribute, and/or modify
+ * it under the terms of the GNU Affero General Public License, version 3
+ * or later ("AGPL"), as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package common
+
+import (
+	"fmt"
+	"net"
+	"net/http"
+	"os"
+	"strings"
+
+	"github.com/urfave/cli/v2"
+
+	"github.com/juicedata/juicefs/pkg/meta"
+	"github.com/juicedata/juicefs/pkg/metric"
+	"github.com/juicedata/juicefs/pkg/object"
+	"github.com/juicedata/juicefs/pkg/utils"
+	"github.com/juicedata/juicefs/pkg/version"
+	"github.com/juicedata/juicefs/pkg/vfs"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
+)
+
+var logger = utils.GetLogger("juicefs")
+
+func CreateStorage(format *meta.Format) (object.ObjectStorage, error) {
+	object.UserAgent = "JuiceFS-" + version.Version()
+	var blob object.ObjectStorage
+	var err error
+	if format.Shards > 1 {
+		blob, err = object.NewSharded(strings.ToLower(format.Storage), format.Bucket, format.AccessKey, format.SecretKey, format.Shards)
+	} else {
+		blob, err = object.CreateStorage(strings.ToLower(format.Storage), format.Bucket, format.AccessKey, format.SecretKey)
+	}
+	if err != nil {
+		return nil, err
+	}
+	blob = object.WithPrefix(blob, format.Name+"/")
+
+	if format.EncryptKey != "" {
+		passphrase := os.Getenv("JFS_RSA_PASSPHRASE")
+		privKey, err := object.ParseRsaPrivateKeyFromPem(format.EncryptKey, passphrase)
+		if err != nil {
+			return nil, fmt.Errorf("load private key: %s", err)
+		}
+		encryptor := object.NewAESEncryptor(object.NewRSAEncryptor(privKey))
+		blob = object.NewEncrypted(blob, encryptor)
+	}
+	return blob, nil
+}
+func ExposeMetrics(m meta.Meta, c *cli.Context) string {
+	var ip, port string
+	//default set
+	ip, port, err := net.SplitHostPort(c.String("metrics"))
+	if err != nil {
+		logger.Fatalf("metrics format error: %v", err)
+	}
+
+	meta.InitMetrics()
+	vfs.InitMetrics()
+	go metric.UpdateMetrics(m)
+	http.Handle("/metrics", promhttp.HandlerFor(
+		prometheus.DefaultGatherer,
+		promhttp.HandlerOpts{
+			// Opt into OpenMetrics to support exemplars.
+			EnableOpenMetrics: true,
+		},
+	))
+	prometheus.MustRegister(prometheus.NewBuildInfoCollector())
+
+	// If not set metrics addr,the port will be auto set
+	if !c.IsSet("metrics") {
+		// If only set consul, ip will auto set
+		if c.IsSet("consul") {
+			ip, err = utils.GetLocalIp(c.String("consul"))
+			if err != nil {
+				logger.Errorf("Get local ip failed: %v", err)
+				return ""
+			}
+		}
+	}
+
+	ln, err := net.Listen("tcp", net.JoinHostPort(ip, port))
+	if err != nil {
+		// Don't try other ports on metrics set but listen failed
+		if c.IsSet("metrics") {
+			logger.Errorf("listen on %s:%s failed: %v", ip, port, err)
+			return ""
+		}
+		// Listen port on 0 will auto listen on a free port
+		ln, err = net.Listen("tcp", net.JoinHostPort(ip, "0"))
+		if err != nil {
+			logger.Errorf("Listen failed: %v", err)
+			return ""
+		}
+	}
+
+	go func() {
+		if err := http.Serve(ln, nil); err != nil {
+			logger.Errorf("Serve for metrics: %s", err)
+		}
+	}()
+
+	metricsAddr := ln.Addr().String()
+	logger.Infof("Prometheus metrics listening on %s", metricsAddr)
+	return metricsAddr
+}
